Fixed a parser issue with terminal order
All of the double terminals needed to be higher priority than their
single character bretheren, for example

'==' needs to be recognized in the table BEFORE '=' or '==' will always
look like two consecutive '=' tokens. which is a syntax error, it
thought. The same applies to all of the two character operators.

I also added != as an alias for <> relational operator, but it's the
same token, just an alias.